Koptions (#15)
* Do something with user_config

Otherwise pyflakes will complain stupidly

* Add -a option

The -a/--add-option flag will allow you to add option to the kernel boot list without needing to copy all of the existing options in again. If you already have the options listed present, kernelstub will ignore them.

* Make add-options more reliable

Split the options to be added into a list and check each item instead of the list as a whole. This prevents duplicate options from being added.

* Update README

* Parse options better

Since options can have (multiple) spaces inside of quotation marks, we need to write a parser to recombine those options into a single list item. Otherwise we may get issues with options becoming split and getting treated individually.

* Create ISSUE_TEMPLATE.md

* Error code changes

If the kernel or initrd images can't be found, we need to exit with 0
so that kernel updates don't fail.

* Fix duplicate version

* Remove flag (#19)

* Error code changes (#12)

If the kernel or initrd images can't be found, we need to exit with 0
so that kernel updates don't fail.

* Add delete option to the parser

* Add kernel-option parser to configuration class

* Implement kernel-option removal

* Fix some bugs

* Pass string version of options where necessary/appropriate

* Breaking API change

* fix help

* Add extra builer file

* Fix accidental tuple in help

Signed-off-by: Ian Santopietro <ian@system76.com>

* Fix for #20

Seems that kernel options provided to the command are not being split into an array, and thus any attempt to iterate on it will iterate by characters instead of words.

Closes #20

* extra space