Add hostname and UUID to entry.
This ensures that each entry applies strictly to the installation that created it.
Before, if two OSs (possibly of different versions) used kernelstub, whichever one last
had kernelstub run would be the one whose entry was written. This change cleans that up
a bit by writing the hostname and the UUID into the filename of the entry. This ensures
that the entry files are unique to that installation, and provides a human-readable way
to select which installation to boot (if the hostname was changed).